--- v2/components/Header/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@snx-v2/Header",
-  "private": true,
-  "main": "index.tsx",
-  "version": "0.0.0",
-  "dependencies": {
-    "@snx-v2/Navigation": "workspace:*",
-    "@synthetixio/contracts-interface": "workspace:*",
-    "react": "^18.2.0"
-  }
-}
