+++ v2/ui/sections/shared/Layout/AppLayout.tsx
@@ -11,7 +11,7 @@ import media from '@snx-v1/media';
 import { delegateWalletState } from 'store/wallet';
 import Header from './Header';
 import SideNav from './SideNav';
-import { Header as V2Header } from '@snx-v2/Header';
+import { Header as V2Header } from '../../../v2-components/Header';
 import useSynthetixQueries from '@synthetixio/queries';
 import Connector from 'containers/Connector';
 import useLocalStorage from 'hooks/useLocalStorage';
